test: Move canvas and editor tests to workflows/editor/ directory
