+++ packages/testing/playwright/tests/ui/workflows/editor/workflow-actions.spec.ts
@@ -7,10 +7,10 @@ import {
 	MANUAL_TRIGGER_NODE_NAME,
 	NOTION_NODE_NAME,
 	SCHEDULE_TRIGGER_NODE_NAME,
-} from '../../config/constants';
-import { test, expect } from '../../fixtures/base';
-import type { n8nPage } from '../../pages/n8nPage';
-import { resolveFromRoot } from '../../utils/path-helper';
+} from '../../../../config/constants';
+import { test, expect } from '../../../../fixtures/base';
+import type { n8nPage } from '../../../../pages/n8nPage';
+import { resolveFromRoot } from '../../../../utils/path-helper';
 
 async function saveWorkflowAndGetId(n8n: n8nPage): Promise<string> {
 	const saveResponsePromise = n8n.page.waitForResponse(
